add: main-page route fix(js): message submit fix(chat/urls): delete excess route
