+++ fchat/apps/chat/urls.py
@@ -1,10 +1,11 @@
 from django.urls import path
-from .views import ChatListView, CreateEventView
+from .views import MainPageTemplateView, ChatListView, CreateEventView
 
 
 app_name = "chat"
 
 urlpatterns = [
+    path('main/', MainPageTemplateView.as_view(), name='main-page'),
     path('<uuid:group_uuid>/', ChatListView.as_view(), name='group'),
     path('add/', CreateEventView.as_view(), name='group-add'),
 ]
