+++ fchat/apps/chat/views.py
@@ -2,7 +2,7 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.decorators import login_required
 from django.urls import reverse
 from django.views.generic.list import ListView
-from django.views.generic import View
+from django.views.generic import View, TemplateView
 from django.forms import ValidationError
 from django.shortcuts import get_object_or_404
 from django.utils.decorators import method_decorator
