+++ fchat/static/src/js/connection.js
@@ -42,3 +42,27 @@ function connectWebSocket() {
 
 // Переподключаемся при изменении URL (если SPA)
 window.addEventListener('popstate', connectWebSocket);
+
+
+const messageInputDom = document.querySelector('#chat-message-input');
+const messageSubmitDom = document.querySelector('#chat-message-submit'); 
+
+messageSubmitDom.onclick = function(e) {
+    sendMessage();
+};
+messageInputDom.addEventListener('keydown', function(e) {
+    if (e.key === 'Enter') {
+        e.preventDefault(); // Предотвращаем стандартное поведение Enter (перенос строки)
+        sendMessage();
+    }
+})
+
+function sendMessage() {
+    const message = messageInputDom.value.trim();
+    // FIXME: ЭТО ЧТО Ж БЛЯТЬ ПОЛУЧАЕТСЯ
+    // ЦИКЛИЧЕСКИЙ ИМПОРТ, СУКА?
+    chatSocket.send(JSON.stringify({
+        'message': message
+    }));
+    messageInputDom.value = '';
+}
